fix process clip
